+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - The **PositionedIndicatorContainer** class is no longer exported from this package, however the source code is available in the example application. 
 - Now the *onRefresh* function will be triggered immediately when the indicator is released in the armed state. Previously, the *onRefresh* function was triggered when the indicator reached a target value in the loading state of `1.0`.
 - Fixed a bug causing the *onRefresh* method not to be triggered on the iOS platform due to bounce physics.
+- Implemented equality operator for *IndicatorStateChange* class.
 - Multiple minor fixes, improvements and optimizations.
 ## 1.2.1
 - Flutter 3.0.0 migration backward compatibility fix ([#31](https://github.com/gonuit/flutter-custom-refresh-indicator/pull/31)) by [Jordan1122](https://github.com/Jordan1122)